KAT: Pipe output of KAT binaries into Python hash checker
The KATs are simply hash-comparisons for the output of the KAT binaries
against hardcoded hashes.

The previous hoisted both the running and the hash computation into
a Python script. This makes it harder to note the simplicity of the KAT,
say if for some reason you need to run things manually.

This commit goes back to the previous model of running the KAT binaries
and feeding their output to a 'hash checker'. It retains the use of Python
instead of the previous bash based pipe chain.
